Audio: improve audio hwsync pts [1/1]
PD#OTT-22384

Problem:
there is a overrun about pts(64bit-->32bit-->64bit) in AudioHal.
the apts is used to 32bit in audiohal,
though recieved 64bit pts form framework.
this 32bit pts will lead an error
when pts from framework is greater than the max value of int type.

Here this error lead to video stuck in tunnel mode.

Solution:
Not use the 32bit apts,
change the 32bit to 64bit in AudioHal.

Verify:
MiBox S4D and Y4.

+++ audio/audio_hal/audio_hwsync_wrap.h
@@ -40,12 +40,12 @@ void aml_hwsync_wrap_set_tsync_resume(audio_hwsync_t *p_hwsync);
 int aml_hwsync_wrap_set_tsync_start_pts(audio_hwsync_t *p_hwsync, uint32_t pts);
 int aml_hwsync_wrap_set_tsync_start_pts64(audio_hwsync_t *p_hwsync,uint64_t pts);
 void aml_hwsync_wrap_set_tsync_stop(audio_hwsync_t *p_hwsync);
-int aml_hwsync_wrap_get_tsync_pts(audio_hwsync_t *p_hwsync, uint32_t *pts);
-int aml_hwsync_wrap_reset_tsync_pcrscr(audio_hwsync_t *p_hwsync, uint32_t pts);
+int aml_hwsync_wrap_get_tsync_pts(audio_hwsync_t *p_hwsync, uint64_t *pts);
+int aml_hwsync_wrap_reset_tsync_pcrscr(audio_hwsync_t *p_hwsync, uint64_t pts);
 bool aml_hwsync_wrap_get_id(void *mediasync, int32_t* id);
 bool aml_hwsync_wrap_set_id(audio_hwsync_t *p_hwsync, uint32_t id);
 bool aml_hwsync_wrap_release(audio_hwsync_t *p_hwsync);
 void aml_hwsync_wrap_wait_video_start(audio_hwsync_t *p_hwsync, uint32_t wait_count);
-void aml_hwsync_wrap_wait_video_drop(audio_hwsync_t *p_hwsync, uint32_t cur_pts, uint32_t wait_count);
+void aml_hwsync_wrap_wait_video_drop(audio_hwsync_t *p_hwsync, uint64_t cur_pts, uint32_t wait_count);
 
 #endif
